resolucao de conflitos review

+++ src/main/java/br/com/fiap/techChallenge3/infraestructure/review/dto/ReviewPublicData.java
@@ -4,6 +4,8 @@ import br.com.fiap.techChallenge3.entity.restaurant.model.Restaurant;
 import br.com.fiap.techChallenge3.entity.review.model.Review;
 import br.com.fiap.techChallenge3.usecase.review.dto.IReviewPublicData;
 
+import java.io.Serializable;
+
 
 public record ReviewPublicData(
         Long id,
@@ -12,7 +14,7 @@ public record ReviewPublicData(
         Double rating,
         String comment) implements IReviewPublicData {
     public ReviewPublicData(Review review){
-        this(   review.getId(),
+        this( (Long) review.getId(),
                 review.getRestaurant(),
                 review.getUserId(),
                 review.getRating(),
